git-svn-id: file:///usr/local/svn/fsc2@7955 934aee4a-8510-0410-ae50-ec5b2ea4bb86

+++ trunk/fsc2/modules/dg2020_util_b.c
@@ -73,7 +73,7 @@ Ticks dg2020_double2ticks( double p_time )
 double dg2020_ticks2double( Ticks ticks )
 {
     fsc2_assert( dg2020.is_timebase );
-    return ( double ) ( dg2020.timebase * ticks );
+    return dg2020.timebase * ticks;
 }
